==> NB00-Geoprivacy-Intro.ipynb <==
# Introduction to Geoprivacy Notebooks

## Requirements

The following Python 3.5+ packages are required to run the notebooks in this case study.
1. Pandas
2. Geopandas
3. Folium and plugins
4. Random
5. Math
6. Numpy
7. Geopy
8. Time
9. H3
10. H3-py

## Approach

1. Start with Cholera use case
2. Kampala MSM (using random points as original)

## References

1. Douglas B. Richardson, Mei-Po Kwan, George Alter & Jean E. McKendry (2015) Replication of scientific research: addressing geoprivacy, confidentiality, and data sharing challenges in geospatial research, Annals of GIS, 21:2, 101-110, DOI: 10.1080/19475683.2015.1027792. URL: https://www.tandfonline.com/doi/full/10.1080/19475683.2015.1027792
